EDSC-4322: Renaming files
