+++ static/src/js/components/AvailableCustomizationsIcons/AvailableCustomizationsIcons.jsx
@@ -12,20 +12,18 @@ import {
 
 import EDSCIcon from '../EDSCIcon/EDSCIcon'
 
-import './CustomizableIcons.scss'
+import './AvailableCustomizationsIcons.scss'
 
 /**
  * Renders icons indicating customization options for access methods.
- * Used by CollectionResultsItem and AccessMethodRadio.
  * @param {boolean} hasSpatialSubsetting
  * @param {boolean} hasVariables
  * @param {boolean} hasTransforms
  * @param {boolean} hasFormats
  * @param {boolean} hasTemporalSubsetting
  * @param {boolean} hasCombine
- * @param {boolean} forAccessMethodRadio - indicates usage for AccessMethodRadio
  */
-export const CustomizableIcons = ({
+export const AvailableCustomizationsTooltipIcons = ({
   hasSpatialSubsetting,
   hasVariables,
   hasTransforms,
@@ -44,7 +42,7 @@ export const CustomizableIcons = ({
     {
       hasSpatialSubsetting && (
         <EDSCIcon
-          className="customizable-icons__icon"
+          className="available-customization-icons__icon"
           title="A white globe icon"
           icon={FaGlobe}
           size="0.675rem"
@@ -54,7 +52,7 @@ export const CustomizableIcons = ({
     {
       hasTemporalSubsetting && (
         <EDSCIcon
-          className="customizable-icons__icon"
+          className="available-customization-icons__icon"
           title="A white clock icon"
           icon={FaClock}
           size="0.675rem"
@@ -64,7 +62,7 @@ export const CustomizableIcons = ({
     {
       hasVariables && (
         <EDSCIcon
-          className="customizable-icons__icon"
+          className="available-customization-icons__icon"
           title="A white tags icon"
           icon={FaTags}
           size="0.675rem"
@@ -74,7 +72,7 @@ export const CustomizableIcons = ({
     {
       hasTransforms && (
         <EDSCIcon
-          className="customizable-icons__icon"
+          className="available-customization-icons__icon"
           title="A white horizontal sliders icon"
           icon={Filter}
           size="0.675rem"
@@ -84,7 +82,7 @@ export const CustomizableIcons = ({
     {
       hasFormats && (
         <EDSCIcon
-          className="customizable-icons__icon"
+          className="available-customization-icons__icon"
           title="A white file icon"
           icon={FileGeneric}
           size="0.675rem"
@@ -94,7 +92,7 @@ export const CustomizableIcons = ({
     {
       hasCombine && (
         <EDSCIcon
-          className="customizable-icons__icon"
+          className="available-customization-icons__icon"
           title="A white cubes icon"
           icon={FaCubes}
           size="0.675rem"
@@ -104,7 +102,7 @@ export const CustomizableIcons = ({
   </>
 )
 
-CustomizableIcons.defaultProps = {
+AvailableCustomizationsTooltipIcons.defaultProps = {
   hasSpatialSubsetting: false,
   hasVariables: false,
   hasTransforms: false,
@@ -113,7 +111,7 @@ CustomizableIcons.defaultProps = {
   hasCombine: false
 }
 
-CustomizableIcons.propTypes = {
+AvailableCustomizationsTooltipIcons.propTypes = {
   hasSpatialSubsetting: PropTypes.bool,
   hasVariables: PropTypes.bool,
   hasTransforms: PropTypes.bool,
@@ -122,4 +120,4 @@ CustomizableIcons.propTypes = {
   hasCombine: PropTypes.bool
 }
 
-export default CustomizableIcons
+export default AvailableCustomizationsTooltipIcons
