+++ static/src/js/components/AvailableCustomizationsIcons/AvailableCustomizationsIcons.scss
@@ -1,4 +1,4 @@
-.customizable-icons {
+.available-customization-icons {
   border-bottom: 1px solid $border-color;
 
   &__icon {
