+++ static/src/js/components/AvailableCustomizationsIcons/AvailableCustomizationsTooltipIcons.jsx
