+++ static/src/js/components/AvailableCustomizationsIcons/AvailableCustomizationsTooltipIcons.scss
