+++ static/src/js/components/AvailableCustomizationsIcons/__tests__/AvailableCustomizationsIcons.test.js
